# Improve performance by removing unnecessary `try catch` block