| Method          | Mean     | Error    | StdDev   | Rank | Gen0   | Allocated |
|---------------- |---------:|---------:|---------:|-----:|-------:|----------:|
| ManualCheck     | 12.28 us | 0.237 us | 0.244 us |    1 | 0.3967 |   3.29 KB |
| TryCatchVisitor | 13.17 us | 0.263 us | 0.270 us |    2 | 0.5035 |   4.12 KB |